Add initial SSL workings (non-active)
This is the changes we worked on during the October CPAN PRC, using
https://github.com/lukas2511/dehydrated to get letsencrypt certs

This change enables static serving of /.well-known/acme-challenge/ from
the /var/www/dehydrated/ dir on the server.

Currently the config settings to make it work are commented out as it
does not fully work yet. The /root/dehydrated/certs/myjudo.net/ path is
install specific so generalising would be good.

Also I needed to chaneg the settings file to serve on 443

Next commits will address these things and get to working SSL support.