README update to explain a little of the DB schema
JJ identified an issue where upon registering DB errors occur, this
I believe was because the latest schema was not applied to the myjudo.db
file in the repo.

I have bumped that, but also put it in the README to make it a little
easier to understand why it breaks. As I am sure I will forget to store
the schema in db/myjudo.db again in the future.

A proper migration feature would solve this too.