Change the param used in /register route to work correctly
Changed the parameter name we pass on the post route to the library.
Make a POST test of the app itself that registers a user.

This test uses the app and the app is writing to the "production"
DB myjudo.db. Which is not good so I need to fix that.