Default techniqu numbers to 0 if not presemt
It seems that mustache does not like the empty technique numbers.
It was generating a large number of warnings and that page would
take a long time to load.

Here we "or" zero, which reduced the warning lines. It's a hack
but seemed to improve the situation till I can look at the
problem properly